week1_project/text_analysis_tool/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
transformers>=4.30.0
nltk>=3.8.0
openai>=1.0.0
torch>=2.0.0
streamlit>=1.25.0
pandas>=2.0.0
matplotlib>=3.7.0